feat: change Queue consumer from EventHub to Executor
