+++ src/core/executor/rpc.ts
@@ -18,8 +18,34 @@ export abstract class RpcExecutor<
     this.dispatcher = new Dispatcher(this.getRepository(), env);
   }
 
-  dispatch(msg: QueueMessage) {
-    return this.dispatcher.dispatch(msg);
+  private async dispatch(msg: Message<QueueMessage>) {
+    return this.dispatcher
+      .dispatch(msg.body)
+      .then((result) => {
+        switch (result) {
+          case "complete":
+          case "ignored":
+          case "misconfigured":
+          case "notfound":
+            msg.ack();
+            break;
+          case "failed":
+            msg.retry();
+            break;
+          default: {
+            const _: never = result;
+          }
+        }
+      })
+      .catch(() => {
+        msg.retry();
+      });
+  }
+
+  async queue(batch: MessageBatch<QueueMessage>) {
+    for (const msg of batch.messages) {
+      await this.dispatch(msg);
+    }
   }
 
   protected abstract getRepository(): Repository;
