+++ src/core/hub/rpc.ts
@@ -2,10 +2,8 @@ import { WorkerEntrypoint } from "cloudflare:workers";
 import * as v from "valibot";
 
 import { type EventHub, EventSink } from ".";
-import type { Executor } from "../executor";
 import type { Repository } from "../repository";
 import type { EventPayload } from "../type";
-import type { QueueMessage } from "./queue";
 import { Config } from "./routing";
 
 const getQueue = (env: Record<string, unknown>) => {
@@ -31,19 +29,6 @@ const getRouteConfig = (env: Record<string, unknown>) => {
   return v.parse(Config, JSON.parse(routing));
 };
 
-const getExecutor = (env: Record<string, unknown>) => {
-  const executor = env.EVENTHUB_EXECUTOR;
-  if (!executor) {
-    throw new Error("cf-eventhub: EVENTHUB_EXECUTOR not set");
-  }
-  if (typeof executor !== "object" || "dispatch" in executor) {
-    throw new Error(
-      "cf-eventhub: value of EVENTHUB_EXECUTOR is not a Executor",
-    );
-  }
-  return executor as Executor;
-};
-
 export abstract class RpcEventHub<
     Env extends Record<string, unknown> = Record<string, unknown>,
   >
@@ -51,48 +36,16 @@ export abstract class RpcEventHub<
   implements EventHub
 {
   private sink: EventSink;
-  private executor: Executor;
 
   constructor(ctx: ExecutionContext, env: Env) {
     super(ctx, env);
     const repo = this.getRepository();
     this.sink = new EventSink(repo, getQueue(env), getRouteConfig(env));
-    this.executor = getExecutor(env);
   }
 
   protected abstract getRepository(): Repository;
 
   putEvent(events: EventPayload[]) {
     return this.sink.putEvent(events);
-  }
-
-  private async dispatch(msg: Message<QueueMessage>) {
-    await this.executor
-      .dispatch(msg.body)
-      .then((result) => {
-        switch (result) {
-          case "complete":
-          case "ignored":
-          case "misconfigured":
-          case "notfound":
-            msg.ack();
-            break;
-          case "failed":
-            msg.retry();
-            break;
-          default: {
-            const _: never = result;
-          }
-        }
-      })
-      .catch(() => {
-        msg.retry();
-      });
-  }
-
-  async queue(batch: MessageBatch<QueueMessage>) {
-    for (const msg of batch.messages) {
-      await this.dispatch(msg);
-    }
   }
 }
